Update patch set 1
Patch Set 1:

> (1 comment)

Afaik both choice and case are a mixin node that does not get stored in datastore so upgrades wouldnt be affected I think. Anyway its master so breakages are expected in the middle of dev cycle so a revert is overreacting tbh, but i restored the old name so downstream doesnt have to change anything: https://git.opendaylight.org/gerrit/#/c/64875/

Brady: consider moving from the deprecated credentials as in 2 releases they would be scheduled for removal anyway.

Patch-set: 1