Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Afaik both choice and case are a mixin node that does not get
 > stored in datastore so upgrades wouldnt be affected I think. Anyway
 > its master so breakages are expected in the middle of dev cycle so
 > a revert is overreacting tbh, but i restored the old name so
 > downstream doesnt have to change anything: https://git.opendaylight.org/gerrit/#/c/64875/
 > 
 > Brady: consider moving from the deprecated credentials as in 2
 > releases they would be scheduled for removal anyway.

It's it's data that's written in a transaction then it gets persisted. If you're unsure if a yang change will break persisted data on upgrade then you should test it to make sure either easy. Re: "its master so breakages are expected" - maybe but how about announcing it ahead of time so folks can prepare or check  ahead of time if other projects will be affected. We really shouldn't just make breaking changes in a vacuum.

Patch-set: 1