Update patch set 6
Patch Set 6: Code-Review+1

> my question above is still unanswered

The problem is not that there were really two NotificationListenerAdapters for JSON stream, because no JSON one created. But it is incorrect to put the NotificationListenerAdapter Object for XML stream to the list for JSON stream.
When WebSocket subcribe JSON stream, you will add the Subcriber with channel to both the NotificationListenerAdapter for XML and the NotificationListenerAdapter for JSON. Is it weird?

Patch-set: 6
Reviewer: Gerrit User 7614 <7614@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1