Minimize yangtools dependencies and bump to 2.0.1
Squash dependencies so that netconf features are pulling in
the bare minimum set of yangtools features directly and use
version ranges for them.

Properties should be avoided and inter-project dependencies should
be expressed via ${upstream}-artifacts dependencyManagement import.

yangtools-2.0.1 contains a fix to multi-revision modules, which are
badly broken in 2.0.0. It also introduces APIs we require to fix
compliance with RFC8040 (i.e. follow RFC7159).

Change-Id: I0e0e63841d511123243560f166db50c2fc9270fb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>