Update patch set 3
Patch Set 3:

> OK, now it only fails in distribution-check with the issue that we
 > just try to fix :)

That doesn't make sense. I specify the version for that package and this change fixed the same failure with the tsdr feature that failed every time w/o it. Are you sure this distribution was build with this patch?  If so then there's another project feature that's making this issue surface. Either way I think this change is beneficial despite the distribution failure.

Patch-set: 3