Convert rfc8040 ListenerAdapter to DOMDataTreeChangeListener
DOMDataChangeListener is scheduled for removal - the bierman02
ListenerAdapter was already converted. Also modified the
ListenerAdapter to use the org.json lib instead of the jackson
lib as the UT revealed that the jackson lib doesn't convert
the XML to JSON correctly and the bierman02 ListenerAdapter
was recently converted to use org.json.

Change-Id: I9e6778a206f412bb8498297cdcb94b77ebba25dd
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>