NETCONF-526 : edit-config payload default-operation needs to be set to default-value
For the PUT/POST calls right now default-operation tag in the edit-config is set to
"none". As per RFC with this value devices can reject a config operation when
elements present in the request whose structure is not already present in the tree.
For example a nested parent-child structure where parent structure does not yet exist
yet in the tree may result in an error based on implementation.
We should  leave this field out ( which will default to merge) and let the device make
the right decision.

Change-Id: I8b85ef072785381897a19724dc94003b8f315bd4
Signed-off-by: Balaji Varadaraju <bvaradar@luminanetworks.com>
(cherry picked from commit 011fbb4e78b64748c4be16c43b3deb46a5c3b23c)