DO NOT MERGE(NETCONF-529): add support for <copy-config> Nitrogen
Nitrogen version of: https://git.opendaylight.org/gerrit/#/c/69605/
for testing.

While <edit-config> is enough to configure any Netconf device,
<copy-config> is more user friendly in scenarios,
when entire configuration has to be written in single request, e.g.
there is no need for explicit delete of nodes that were present in DS.

This patch adds support for <copy-config> to mdsal-netconf-connector.
Only <candidate/> is supported.

To be included in subsequent patches:
- move common test code to AbstractNetconfOperationTest
- move common edit-config and copy-config code to AbstractEdit class

Change-Id: I4eb2d1d1c82e60acc1502b01265de5ec3aab4cd8
Signed-off-by: Marek Gradzki <mgradzki@cisco.com>