Update patch set 3
Patch Set 3:

The restconf is not doing that anymore as that was creating many other issues with real devices. This fix provides us flexibility  as we are delegating this to device to make the right decision. This is already merged in Carbon and tested with multiple commercial software/hardware devices.

Patch-set: 3
Label: Verified=0