Update patch set 3
Patch Set 3: Code-Review-1

I've just done the review for this on master branch, so I guess the comments from that review apply also here.

For the future, can we first fix the master branch and after that cherry=pick the patches to other stable branches if needed? This is really hard to follow..

Patch-set: 3