Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: Ia1018cd2a883c6a1622b0ab817076ea1aba19c49
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>