Allow specification of ODL's HELLO message
Some devices are very particular about what they will accept
for a HELLO message.  This change allows specification of a custom
message per device mount.  If one is not specified, then the
existing default is utilized.

Additionally, certain tests were using very out of date versions of
netconf-node-topology.  This is due to the fact that the test
yang files are located in src/test/resources, and were out of date
with the netconf-node-topology used in src/main/yang.  These were
updated to the most up-to-date version as of this patch.

JIRA: NETCONF-543
Change-Id: I4049a87a7781655d8d396bb35eb43aab2eca0223
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>