Specify necessary dependencies
odl-restconf-common relies on AAA bringing in Jersey 1.X dependencies.
Relying on transitive dependencies is not good practice, since it
binds the upstream to providing them continuously.  AAA is undergoing
cleanup related to Jersey implementation, part of which is an experiment
to switch on Jersey 2.  In order to do this, we remove the Jersey 1
"com.sun.jersey" dependencies.  This causes RESTCONF to have missing
runtime dependencies.  This patch fixes this issue properly.

Change-Id: Idb32f44573684770b9a57675dfadf010d8a16c9e
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>