Use ControllerContext non-statically
This is a preliminary patch for removing the web.xml and converting
to the new programmtic web API which will allow us to remove the
static instances that bridge between the web-instantiated
and blueprint-instanriated components. First up is the
ControllerContext - this patch keeps the static instance temporarily
for the RestconfApplication but converts the rest of the code to
inject the ControllerContext instance and reference it non-statically.

This patch touches a lot of files but most are UTs.

Change-Id: I08e2ee24e518c01bb1d5fdb5796abd9354dccfa2
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>