Remove static TransactionChainHandler instance
RestConnectorProvider maintained a static instance in order
to reset the DOMTransactionChain statically. The reset logic
was moved to TransactionChainHandler and call sites in the
static Utils were modified to pass the TransactionChainHandler via
the TransactionVarsWrapper and call reset. This allows the static
instance to be removed. The TransactionChainHandler singleton is now
created via blueprint.

Change-Id: I483ac83103286146bc636b7f1e6ba0df70f2f5a6
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>