Remove RestConnectorProvider
RestConnectorProvider has been reduced to just initialize the
ServiceWrapper instance so is no longer needed. Also the ServiceWrapper
interface is no longer needed and, as a result, the ServicesWrapperImpl
class was renamed to just ServiceWrappers. The ServiceWrappers
is now instantiated/initiated via blueprint. The static instance
remains temporarily for the RestconfApplication.

Change-Id: I643d18a6980b5f51262396c49af281515b94868f
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>