Remove controller DCL-related references
DataChangeListener, DataChangeScope and friends are being removed
from the controller. The DataChangeScope enum is part of the NB data
change stream registration API although the implementation was removed
in Oxygen when it was internally changed to use DTCL. We may still
want/need to support the scope on top of DTCL so the enum was copied
to restconf-common.

Change-Id: If2ebe7b863975b4cae63c838d60f83b33e9e37b4
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>