Make RuntimeRpc revision aware
Do not use SchemaContext.findModule(URI),
which looks for a module instance with specified namespace
and no revision.
Such lookup fails for modules with revision defined.

Instead (as already done in AbstractEdit)
use SchemaContext.findModules and use the newest module
if more than one module with the same namespace is found.

Change-Id: I7b6bdcb9156669645e19a6395e1b74995328154d
Signed-off-by: Marek Gradzki <mgradzki@cisco.com>