Unit test for Test-optiont
Test option is currently ignored by ODL Netconf.
One can specify test-then-set when validate is not advertsed.
Validate is done anyway when XML is parsed, later also by Yangtools.

And according to RFC minimum is checking for syntax errors.
so it would be NOOP anyway.

We don't have really a usecase which requires test-option,
no changes are required to current code.

Change-Id: I75055a3311de2dcb8b17e81db7aa7e009f25681b
Signed-off-by: Marek Gradzki <mgradzki@cisco.com>