NECONF-524 : Setting the netconf keepalive logic to be more proactive.
When a NETCONF device loses connection with the controller in exceptional
conditions such as the management interface going down, underlying TCP
connection would not be closed as there will be no fin packet exchange or
reset. At this point NETCONF state will be set as connected until the
underlying OS determines the connection is stale which can happen after
considerable amount of time. This fix makes the Keepalive packets more
proactive to determine such conditions.This fix also sets the listener
properly which is required for this fix as well.

Change-Id: Ia91a057c32e589adde676f5fa10e59fd767cc0be
Signed-off-by: Balaji Varadaraju <bvaradar@luminanetworks.com>