Keepalive thread fails during long running Transactions causing device disconnects
During long netconf transactions, if a device is not able to respond to keepalive rpc,
the keepalive thread disconnects from device and marks it offline causing the disruption.
Idle state handler on netty channel will prevent the keepalive from running if the channel is
active. In other words, the keepalive rpc will be sent only if the channel has been idle for
time.

Change-Id: I7bf9f1f532c344c6a105a6374e0968679119adec
Signed-off-by: Atul Gosain <agosain@brocade.com>