Keepalive causes disconnects during long running Tx
During long netconf transactions, if a device is not able
to respond to keepalive rpc, the keepalive thread
disconnects from device and marks it offline causing the
disruption. Idle state handler on netty channel will prevent
 the keepalive from running if the channel is active.

In other words, the keepalive rpc will be sent
only if the channel has been idle for some time.

Change-Id: I0a19c50235b7cea793f590dd29ed20fdb9be53f0
Signed-off-by: Atul Gosain <agosain@luminanetworks.com>