NETCONF-557: Add support for URL capability
This patch brings support of URL capability
in <edit-config> and <copy-config> RPCs
to mdsal-netconf-connector.

Remote config upload is not supported.
Remote to remote operations are not supported.

The capability is advertised as:

urn:ietf:params:netconf:capability:url:1.0?scheme=file

but config download is also supported via http and https.

Change-Id: Idb5bae4e24ff2a098bc60bf24c36fb40113fd8f5
Signed-off-by: Marek Gradzki <mgradzki@cisco.com>