Update patch set 1
Patch Set 1:

More specifically, we need calls to invokeRpc (which seems to be supported in JSONRestconfService) and subscribeToStream (which is not). It's true that RestconfService drags with it some internal stuff but the functionality is already exposed. 

This patch is what I intended to use this: https://git.opendaylight.org/gerrit/#/c/77273/

Patch-set: 1
Reviewer: Gerrit User 5357 <5357@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0