Specialize protocol-framework to netconf
As a first step towards sanity, we eliminate protocol-framwork
by moving it into netconf-api, but specialized to deal with
NETCONF protocol only.

This eliminates one level of abstraction, allowing us to reduce
the number of classes and artifacts involved. We also sacrifice
tests, as most of the mechanics should be covered by netconf-level
tests.

JIRA: NETCONF-554
Change-Id: Iba41c5ad72a0148d4e1d768d3122ddebce18f4c2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>