Update patch set 2
Patch Set 2: Code-Review-1

> (1 comment)

Rpcs always fail with: https://jira.opendaylight.org/browse/NETCONF-593
so Im unable to test this. Will have to look at that for now. However according to the unit tests this only introduces the namespace in rpc output, but in the element thats after the output element.
According to the rfc the namespace should be present in the output element so to track this I opened https://jira.opendaylight.org/browse/NETCONF-592

Patch-set: 2
Label: Code-Review=-1