Rework NetconfUtil.writeNormalizedNode()
This refactors existing methods to support writing raw data into
either a Document or an Element, without exposing JAXP details, like
XMLStreamException and DOMResult to the callers.

Also create a specialized createAnyxmlNode, which creates an
element bound to specified document and fills it with specified
normalized node and encapsulates it into an AnyXmlNode.

JIRA: NETCONF-603
Change-Id: Icff5e045c4e901bb53990177d448a1ea43952c62
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>