Initialize EXISchema grammars lazily
Creating schema-backed netconf grammar is costly operation. Do not
create it until we actually need it.

This way we do not introduce any performance regression for netconf
sessions that do not want to use schema-informed grammars for EXI
encoding.

Also do not force grammar instantiation when a mismatching option
is specified by the peer.

JIRA: NETCONF-584
Change-Id: I83f5e18f423cd21c1dc993980f42e030f36a12ed
Signed-off-by: Jakub Morvay <jmorvay@frinx.io>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>