Update patch set 1
Patch Set 1: Code-Review-1

This is a change in behavior and will not work with servers which configure *only* simple password-interactive authentication.

Hence the order should be public-key,password,keyboard-interactive -- if the last fails because of it being something we do not handle, it is probably that can be fixed.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1