WIP: Replace EditOperationNormalizedNodeStreamWriter
Since NormalizedNodes no longer hold attributes directly, but these
are rather a separate structure communicated through an extension
to NormalizedNodeStreamWriter, the entire implementation of how
edit-config is split into individual changes needs to be reworked.

This patch replaces EditOperationNormalizedNodeStreamWriter and
DataTreeChangeTracker with SplittingNormalizedNodeMetadataStreamWriter,
which is the primary receiver of events. It takes care of delegating
them to an ImmutableNormalizedNodeStreamWriter, subclassed as
ComponentNormalizedNodeStreamWriter, which exposes build() operation,
which peels off the current builder and builds it.

Since ModifyAction hosted in yangtools was only used as an attribute
in NormalizedNodes and hence was removed, it is introduced here
in netconf-api.

The sal-netconf-connector part of this change takes care of emitting
the ModifyAction request in terms of NormalizedNodeMetadataStreamWriter.

Change-Id: I83affe10d1334ecebc9fafb9c76021c2de389239
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>