RPC fails on Netconf clustered topology
Some of the type of messages fail on clustered topology due to
deserialization issues with these message classes. The proxy used in
these classes prevent it from being deserialized by akka infra.
Removing this extra layer which shouldnt be required to use
externalizable on these messages.

Change-Id: I9ce8b262a4718a202952f81cce03f4bdfc8945cd
Signed-off-by: agosain <agosain@luminanetworks.com>

Casting to collection of RpcErrors

Change-Id: I0ed5a895ba44128b9fed7e520081f740d5f12dba
Signed-off-by: agosain <agosain@luminanetworks.com>