Convert anyxml nodes lazily
The problem is talking about the direct creating of ContainerNode
from the Anyxml in the NetconfMessageTransformer. Later, while
the YT is parsing data, it expects the anyxml on that place
according to the schema context. In this way, we need to use
the AnyXml builder while the NetconfMessage is processing.

Transformation to a container-based layout is performed afterwards
as needed.

JIRA: NETCONF-616
Change-Id: Iab00b9ef712dbebf0661546550b10bf04507e97a
Signed-off-by: Jakub Toth <jtoth@luminanetworks.com>
Signed-off-by: Miroslav Macko <mmacko@luminanetworks.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>