NETCONF-616 Yangtools is not processing the output of get-config RPC in ietf-netconf YANG model
The problem is talking about the direct creating of ContainerNode
from the Anyxml in the NetconfMessageTransformer. Later, while
the YT is parsing data, it expects the anyxml on that place
according to the schema context. In this way, we need to use
the AnyXml builder while the NetconfMessage is processing.
We can transform it into normalized nodes later.

Change-Id: Iab00b9ef712dbebf0661546550b10bf04507e97a
Signed-off-by: Jakub Toth <jtoth@luminanetworks.com>
Signed-off-by: Miroslav Macko <mmacko@luminanetworks.com>