Teach NETCONF about YANG 1.1 actions in cluster topology
NETCONF allows invocation of actions but only in Non-cluster
topology which works absolutly fine but when invoked in cluster
topology it failed, usecase was not supported.

This patch provides support for Yang 1.1 action in cluster topology
NETCONF code base is updated to support action, can be invoked in
cluster topology.

JIRA: NETCONF-635
Change-Id: Id7f5c9ef4a515c28d5da8bab29724bb353c10d13
Signed-off-by: ajay_dp001 <ajay.deep.singh@est.tech>