Update patch set 8
Patch Set 8: Code-Review-1

> If I'm not mistaking, this is fixing bug 4548: https://bugs.opendaylight.org/show_bug.cgi?id=4548
 > 
 > If it is, please update the commit message to reflect that, and the
 > bug with this patch.
 > Thanks

Actually this patch was originally intended for : https://bugs.opendaylight.org/show_bug.cgi?id=5634

It can also be used as a more generic solution for 4548 but I'll just change 4548 as depending on 5634.
Nevertheless you should but the 5634 title and number in the first line of the commit instead of the existing first line.

Patch-set: 8
Label: Code-Review=-1