Update patch set 10
Patch Set 10:

Also just an fyi but you could have used the existing netconf-cli as base and extended or changed it and you would have ended up with a generic netconf cli that you could have used with netconf for mdsal as an mdsal cli. It still would need a lot of work though so this is a faster solution.

Patch-set: 10