Update patch set 8
Patch Set 8:

> Couple of questions:
 > 
 > 1) which revision is implemented here? call home is in AUTH48 to
 > being RFC8071: https://www.rfc-editor.org/auth48/C279

We started implementation using the ietf-draft-call-home-08 spec. I think we should move
forward to the newer drafts as the devices support the newer behaviors.

 > 2) For configuration, would it be possible to use the definitions
 > from
 > https://tools.ietf.org/html/draft-ietf-netconf-netconf-client-server-01
 > ?

This is a very reasonable request. We will create a high priority story to implement in
our next round of development. However we do not have the funding to make these
changes before the Carbon release. Perhaps we can set the expectations correctly
by marking this as an experimental feature?

Patch-set: 8