Update patch set 8
Patch Set 8:

(17 comments)

> Couple of questions:
 > 
 > 1) which revision is implemented here? call home is in AUTH48 to
 > being RFC8071: https://www.rfc-editor.org/auth48/C279
 > 

We started with draft-ietf-netconf-call-home-08. We'd like to move forward to support newer drafts as the devices themselves advance.

 > 2) For configuration, would it be possible to use the definitions
 > from
 > https://tools.ietf.org/html/draft-ietf-netconf-netconf-client-server-01
 > ?

This is a very reasonable request. We will create a story to move to these
models in our next round of development, but we do not have the funding
to do this before the Carbon release. Perhaps the call home feature can be
labeled as experimental to set the expectation that the API will change.

Patch-set: 8
Label: Verified=0