Update patch set 1
Patch Set 1:

Added deleteSessionOnDuplicateSshKey() to the CallHomeProtocolSessionContext.java, because the server details are not stored in callHomeProvider unless the SSH key is verified to begin with.  With the deleteSessionOnDuplicateSshKey, we are able to close the ssh session, if there is a duplicate ssh host key.

Patch-set: 1
Reviewer: Gerrit User 8641 <8641@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0