Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Jamo, I believe we do not have any RFC 8040 CSIT yet, is it possible yo verify this fix manually?
> 
> We do have some very basic rfc8040 CSIT with GET requests only, but for sure nothing to test depth.
> 
> I did test locally with master branch though, and I don't see the bug. I've replied via email and
> we can continue the discussion there. Also, maybe it's better to open a JIRA to help track what's
> going on?
> 
> email discussion:
> https://lists.opendaylight.org/g/netconf-dev/topic/support_query_at_different/71562437

Actually, I did not fully understand the original issue. The bug is that we don't
get the yang module name in the output, not that depth itself is not working. I replied
to the email and tested this fix locally. Thanks for the work.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1