Factor out SchemaResourceManager
Management of schema cache on the filesystem is copy&pasted in two
places, which creates a compatibility problem. Furthermore current
code relies on static wiring, which is going to be non-workable
very soon.

Refactor AbstractTopology to split out schema management into a
separate component, which is injected into both AbstractTopology
subclasses and clustered topology.

JIRA: NETCONF-672
Change-Id: If1284a08f9525a3396f2d39e2a4399366edee7ae
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>