Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Jamo, do we still need to wait?

No, I think we are ok to get this merged now. I'm not a committer on netconf, so I cannot
merge it. I think Robert might be the only active committer, so we just wait for him to
get to this.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>