Update patch set 1
Patch Set 1:

This may be the totally wrong approach, but it does "fix" the things I was hoping for in NETCONF-666. If it's even close, I can iterate on it to get it right. If it's garbage, I can try something new but
will need guidance. If it's way over my head, I'll have to punt to someone more expert.

BTW, master (aluminium) does not work right now, so to test the distribution created with this
patch it needs to be done via stable/magnesium which is what I'm doing locally.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>