Introduce parent/child relationship in fields query to resolve NETCONF-660
In setting up the fields query, the current code sets up an index of hashes, one for every depth of the search query. It keeps track of where it is in the query by a variable called currentDepth. The current bug was a premature termination of the search when a sub-field query was present.

However, an extended set of test cases, revealed that there were other issues with how the HashIndex was setup. In particular, by splitting the query purely based on where the search fields are based from a depth perspective, it loses the parent/child relationship. This manifested itself when a child with the duplicate name appeared in two different sibling containers. The query in this case, depending on where it falls in the order of query, would return the wrong value.

This code change proposes adding that parent/child relationship when building up the HashIndex for the query. That way at any depth in the query, a child knows who its parent is, and based on the query whether the child needs to be written to the output stream or not.

Signed-off-by: santanude <santanu.de@xoriant.com>
Change-Id: I379490400ce5950a693a7a811eb4a0a54bd9d975