Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 1: Code-Review-1
> > 
> > (3 comments)
> 
> Done, but I do not know how to deal with this UT.

I commented in a different branch of this patch, but wondering if the UT is valid or
not. but I think commenting it out is what is not right. If it's not valid and we need
to remove it, then just delete all the lines and probably add a note in the commit
message to why it was removed. If it's still valid, you need to update it so that it
passes (assuming it's failing now and that's why you commented it out)

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>