Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> I do not believe a straight replacement is appropriate. Doesn't that force existing users to change their code?

Yes this is one problem (probably will be better talk about this in the next kernel call?). 
We see there has in past implemented different types of communication as in RFC(8040, Bierman). Meaning behind this is to start using what is in RFC. 

There is no problem having both types at the same time, this is possible. Problem is on side interfaces: 
rests/data/ietf-restconf-monitoring:restconf-state/streams/*  this interface now every time returns a link to a web service. But in RFC there is link to SSE 

<streams xmlns="urn:ietf:params:xml:ns:yang:ietf-restconf-monitoring">
         <stream>
            <name>NETCONF</name>
            <description>default NETCONF event stream</description>
            <replay-support>true</replay-support>
            <replay-log-creation-time>\
               2007-07-08T00:00:00Z\
            </replay-log-creation-time>
            <access>
               <encoding>xml</encoding>
               <location>https://example.com/streams/NETCONF\
               </location>
            </access>
            <access>
               <encoding>json</encoding>
               <location>https://example.com/streams/NETCONF-JSON\
               </location>
            </access>
         </stream>
</streams>

Due to there is several possibilities but no one is best :
1)prepare for WS some new interface => problem for users because link is changed
2)have different interface for SSE =>  this is against RFC
3)somehow update response to have both link here => have both problem as 1) and 2) plus it can brake validation of response 

This change is now only for restconf8040 where we hope it is used only small amount of users and this change will be merges in Aluminum (main version)

What do you think about this? What will be best?

Patch-set: 5