Update patch set 1
Patch Set 1:

Kailash,

one of the verify jobs for these ODL java projects is called distribution-check. Look in the
console logs of that job and you'll find something called BUNDLE_URL that will point you to
the full distribution that this patch created. You can take that full URL and pass it in to
a CSIT job to test how it looks. I know this is just cosmetic, but you can check it out with
this since I know you are also working in the sandbox too.

Additionally, these jobs can take a gerrit keyword trigger to run CSIT. see this file:
  https://github.com/opendaylight/releng-builder/blob/master/jjb/netconf/netconf-patch-test.yaml

you would type a gerrit comment in the format like: test-{project}-{feature}. so a single comment
of "test-netconf-core" will run two jobs you can see in that .yaml file:
  netconf-csit-1node-gate-userfeatures-all-{stream}
  netconf-csit-1node-gate-callhome-only-{stream}

That stream variable is taken care of and will be the branch of this patch. so master for this
one which gets mapped to Aluminium for csit job names.

all just FYI.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>